pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quicksight-gen"
version = "0.1.0"
description = "Programmatic AWS QuickSight analysis generator for financial reporting"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "boto3-stubs[quicksight]>=1.34",
]

[project.scripts]
quicksight-gen = "quicksight_gen.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
